metadata-display // added metadata display, tweaked experiment data, tweaked composition and dir structure of components
