+++ src/experiments/cube_plus/index.js
@@ -88,6 +88,10 @@ export const AnotherCubeExperiment = () => {
 
 export default {
   id: 'cube_plus',
-  name: 'Another Cube',
+  metadata:{
+    name: 'Another Cube',
+    author: 'JoskerDu',
+    description: '',
+  },
   component: AnotherCubeExperiment,
 };
