+++ src/experiments/index.js
@@ -3,7 +3,7 @@ import simpleCube from './simple_cube/';
 import anotherCube from './cube_plus/';
 import turrelCube from './cube_turrel/';
 
-// Object to be used for route resolution
+// Object to be used for route resolution and metadata
 const experiments = [simpleCube, anotherCube, turrelCube];
 
 export default experiments;
